+++ tests/arm/checks/resource/test_AKSNetworkPolicy.py
@@ -17,7 +17,7 @@ class TestAKSNetworkPolicy(unittest.TestCase):
         summary = report.get_summary()
 
         self.assertEqual(summary['passed'], 1)
-        self.assertEqual(summary['failed'], 4)
+        self.assertEqual(summary['failed'], 1)
         self.assertEqual(summary['skipped'], 0)
         self.assertEqual(summary['parsing_errors'], 0)